More package restructuring changes
