+++ packages/graphics/src/core/ImageLoader.cpp
@@ -1,2 +1,2 @@
 #define STB_IMAGE_IMPLEMENTATION
-#include <graphics/textures/ImageLoader.h>
+#include <graphics/core/ImageLoader.h>
