+++ packages/graphics/src/core/shaders/Shader.cpp
@@ -1,5 +1,5 @@
-#include <graphics/shaders/Shader.h>
-#include <graphics/buffers/Buffer.h>
+#include <graphics/core/shaders/Shader.h>
+#include <graphics/core/Buffer.h>
 #include <graphics/textures/Texture.h>
 #include <math/Matrix4.h>
 #include <common/Assertions.h>
