+++ packages/sampleapp/src/main.cpp
@@ -1,12 +1,12 @@
 #include <graphics/Engine.h>
-#include <graphics/windows/WindowManager.h>
-#include <graphics/windows/Window.h>
-#include <graphics/windows/InputController.h>
-#include <graphics/buffers/Buffer.h>
+#include <graphics/core/windows/WindowManager.h>
+#include <graphics/core/windows/Window.h>
+#include <graphics/core/windows/InputController.h>
+#include <graphics/core/Buffer.h>
 #include <graphics/textures/Texture.h>
 #include <graphics/textures/TextureManager.h>
-#include <graphics/shaders/Shader.h>
-#include <graphics/shaders/ShaderManager.h>
+#include <graphics/core/shaders/Shader.h>
+#include <graphics/core/shaders/ShaderManager.h>
 #include <graphics/geometry/BoxGeometry.h>
 #include <graphics/entities/Mesh.h>
 #include <graphics/materials/MeshMaterial.h>
